chore(x-schemas): Moving them to account for error on OpenRPC Spec
They are now in `#/x-schemas` and not `#/components/x-schemas`
